Update pricing page (ARE-100)
